# TIMX 543 - Optimize current records metadata queries and data retreival