### Purpose and background context

This PR makes some improvements to metadata queries and data retrieval for "current" records, recalling that a "current" record is the most recent version of a `timdex_record_id` in the dataset.

While working on a Marimo notebook that was making quite a few queries against `metadata.current_records`, there was often a 8-10 second delay before results started streaming in.  This was mostly due to how the `TIMDEXDatasetMetadata` class constructed DuckDB tables and views for current records.

The most impactful change is the decision to change the DuckDB asset `metadata.current_records` from a **view** into a [**temporary table**](https://duckdb.org/docs/stable/sql/statements/create_table.html#temporary-tables).  This is implemented by a [rewrite of `TIMDEXDatasetMetadata. _create_current_records_view()`](https://github.com/MITLibraries/timdex-dataset-api/pull/168/commits/25d443019b03af503d99857264f60ca0fad8aacf#diff-248bd5777a91fa90faf6756abf55873d051961692822f40d2db57f11c147671aL416-L468).

By materializing the "current" records metadata to a temporary table -- which actively loads into memory, but will spill to disk if memory constraints are reached -- we get orders of magnitude faster current records queries.

We may need  to revisit this approach if we find ourselves with 25-50 million "current" records, but that is likely a long ways off in the TIMDEX ecosystem.

**NOTE**: it feels worth mentioning one of the strengths of the current DuckDB context approach.  This entire change was 100% DuckDB SQL.  There are relatively no moving parts in python code or data retrieval needed, this was just some refactoring of SQL tables, views, and approaches.  While I'm confident this is an _improvement_ to the current "current" records approach, it does not prohibit us from future improvements and changes. 

The remaining commits do a little cleanup of method naming, and add a `LIMIT` clause when reading records, but they are pretty minor in comparison.

### How can a reviewer manually see the effects of these changes?

1- Set Dev1 AWS `TimdexManagers` credentials in terminal and set env vars:
```shell
TDA_LOG_LEVEL=DEBUG
WARNING_ONLY_LOGGERS=asyncio,botocore,urllib3,s3transfer,boto3,MARKDOWN
TIMDEX_DATASET_LOCATION=s3://timdex-extract-dev-222053980223/dataset_scratch/prod-clone
```

2- Open Ipython shell
```shell
pipenv run ipython
```

3- Load a `TIMDEXDataset` instance:
```python
import os

from timdex_dataset_api import TIMDEXDataset
from timdex_dataset_api.config import configure_dev_logger

configure_dev_logger()

td = TIMDEXDataset(os.environ["TIMDEX_DATASET_LOCATION"])
```

Note that the initial load is a bit slower, but still quite reasonable for connecting to a remote data lake architecture.  We may want to explore the ability to load a `TIMDEXDataset` instance _without_ some of these read capabilities, e.g. when only performing writes (e.g. Transmogrifier), but that feels like an optimiziation for a later time.

4- Perform a metadatda query that formerly took anywhere from 15-20 seconds:
```python
df = td.conn.query("""select * from metadata.current_records where source='alma' and action='index';""").to_df()
print(len(df))
#  3960267
```

5- Retrieve data that relies on current records (can `ctrl + c` to kill early):
```python
for _ in td.read_batches_iter(table='current_records', columns=['timdex_record_id','transformed_record'], source='alma', action='index'):
    pass
# ...
# DEBUG:timdex_dataset_api.dataset:Metadata query identified 3960267 rows, across 213 parquet files, elapsed: 4.23s
# ...
# ...
```

The thing to note here is the metadata query is quick, about 4.23 seconds.  With that quick and complete, we can move into yielding data quicker.

6- Lastly, the 'ole needle-in-a-haystack query to demonstrate the real benefit of this query:
```python
df = td.read_dataframe(table='current_records', source='libguides', action='index')
# ...
# DEBUG:timdex_dataset_api.dataset:Metadata query identified 372 rows, across 64 parquet files, elapsed: 0.04s
# DEBUG:timdex_dataset_api.dataset:read_batches_iter batch 1, yielded: 372 @ 68 records/second, total yielded: 372
# DEBUG:timdex_dataset_api.dataset:read_batches_iter() elapsed: 5.48s
```

Two things to note here:
- a _very_ quick metadata query, around 0.04s
- a _very_ quick data retrieval, even across 64 distinct parquet files

### Includes new or updated dependencies?
NO

### Changes expectations for external applications?
YES: faster current records querying and data retreival!

### What are the relevant tickets?
- https://mitlibraries.atlassian.net/browse/TIMX-543